Reactivate all text editor tests
Due to the Monaco uplift, we had to temporarily skip a test, because
it affected the default preference regarding auto-save.

To re-enable the test, we now explicitly set the auto save preference to
`off` before the test. This way the preference value is ensured to be
consistent.

See also https://github.com/eclipse-theia/theia/pull/10736

Fixes https://github.com/eclipse-theia/theia/issues/10891

Change-Id: I6ab22a71848e174313a30f9121cb4b3dec363f12